+++ resources/views/livewire/invoice-lines.blade.php
@@ -1,20 +1,20 @@
 <div>
     <h6 class="text-uppercase section-h6-title"><span class="fa fa-cart-arrow-down" aria-hidden="true"></span> {{ ucfirst(__('laravel-crm::lang.invoice_lines')) }} <span class="float-right"><button class="btn btn-outline-secondary btn-sm" wire:click.prevent="add({{ $i }})"><span class="fa fa-plus" aria-hidden="true"></span></button></span></h6>
     <hr class="mb-0" />
-    <script type="text/javascript">
+    {{--<script type="text/javascript">
         let products =  {!! \VentureDrake\LaravelCrm\Http\Helpers\AutoComplete\products() !!}
-    </script>
+    </script>--}}
     <span id="invoiceLines">
         <div class="table-responsive">
-            <table class="table">
-                <thead>
+            <table class="table table-sm table-items">
+                {{--<thead>
                     <tr>
                         <th scope="col" class="border-0">{{ ucfirst(__('laravel-crm::lang.name')) }}</th>
                         <th scope="col" class="col-3 border-0">{{ ucfirst(__('laravel-crm::lang.price')) }}</th>
                         <th scope="col" class="col-2 border-0">{{ ucfirst(__('laravel-crm::lang.quantity')) }}</th>
                         <th scope="col" class="col-3 border-0">{{ ucfirst(__('laravel-crm::lang.amount')) }}</th>
                     </tr>
-                </thead>
+                </thead>--}}
                 <tbody>
                 @foreach($inputs as $key => $value)
                     @include('laravel-crm::invoice-lines.partials.fields')
@@ -22,6 +22,16 @@
                 </tbody>
                 <tfoot id="invoiceLinesTotals" class="tfoot">
                     <tr>
+                        <td></td>
+                        <td></td>
+                        <td></td>
+                        <td></td>
+                        <td class="text-right align-middle">
+                            <button class="btn btn-outline-secondary btn-sm" wire:click.prevent="add({{ $i }})"><span class="fa fa-plus" aria-hidden="true"></span></button>
+                        </td>
+                    </tr>
+                    <tr>
+                        <td></td>
                         <td></td>
                         <td></td>
                         <td class="text-right align-middle">{{  ucfirst(__('laravel-crm::lang.sub_total')) }}</td>
@@ -43,6 +53,7 @@
                 <tr>
                     <td></td>
                     <td></td>
+                    <td></td>
                     <td class="text-right align-middle">{{  ucfirst(__('laravel-crm::lang.tax')) }}</td>
                     <td>
                      @include('laravel-crm::partials.form.text',[
@@ -61,6 +72,7 @@
                  <tr>
                     <td></td>
                     <td></td>
+                     <td></td>
                     <td class="text-right align-middle">{{  ucfirst(__('laravel-crm::lang.total')) }}</td>
                     <td>
                    @include('laravel-crm::partials.form.text',[
@@ -84,7 +96,7 @@
     @push('livewire-js')
         <script>
             $(document).ready(function () {
-                $(document).delegate("input[name^='invoiceLines']", "focus", function() {
+                /*$(document).delegate("input[name^='invoiceLines']", "focus", function() {
                     var number = $(this).attr('value')
                     $(this).autocomplete({
                         source: products,
@@ -96,7 +108,22 @@
                         highlightClass: 'text-danger',
                         treshold: 2,
                     });
-                })
+                })*/
+
+                window.addEventListener('addedItem', event => {
+                    $("tr[data-number='" + event.detail.id + "'] select[name^='invoiceLines']").select2()
+                        .on('change', function (e) {
+                            @this.set('product_id.' + $(this).data('value'), $(this).val());
+                            @this.set('name.' + $(this).data('value'), $(this).find("option:selected").text());
+                            Livewire.emit('loadInvoiceLineDefault', $(this).data('value'))
+                        });
+                });
+
+                $("select[name^='invoiceLines']").on('change', function (e) {
+                    @this.set('product_id.' + $(this).data('value'), $(this).val());
+                    @this.set('name.' + $(this).data('value'), $(this).find("option:selected").text());
+                    Livewire.emit('loadInvoiceLineDefault', $(this).data('value'))
+                });
             });
         </script>
     @endpush